+++ Source/Core/InputCommon/ControllerInterface/Device.h
@@ -7,10 +7,10 @@
 #include <string>
 #include <vector>
 
-#include "Common/Common.h"
+#include "Common/CommonTypes.h"
 
 // idk in case I wanted to change it to double or something, idk what's best
-typedef float ControlState;
+typedef double ControlState;
 
 namespace ciface
 {
